Again Some Of My New Learnings

+++ calc1.py
@@ -1,6 +1,6 @@
 choice = input("Choose one:\n 1. Addition\n 2. Subtraction\n 3. Multiplication\n 4. Division\n")
-num1 = int(input("Enter the first number:\n"))
-num2 = int(input("Enter the second number:\n"))
+num1 = float(input("Enter the first number:\n"))
+num2 = float(input("Enter the second number:\n"))
 
 if choice == "1":
     print(num1,"+", num2, "=", num1 + num2)
